+++ CHANGES.md
@@ -1,5 +1,10 @@
 # CHANGES to json-schema-to-zodexy
 
+## 0.5.0
+
+- feat: support $id, title, deprecated conversion
+- fix: change output of description to meta.description as per recommended Zod behavior
+
 ## 0.4.0
 
 - refactor: switch to nodenext; drop default export